fix: refactor link section files
